requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.10.0
branca==0.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.4
django-appconf==1.0.5
django-compressor==4.1
djangorestframework==3.14.0
flake8==6.0.0
folium==0.13.0
idna==3.4
isort==5.10.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.5
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.6.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.6
rcssmin==1.1.0
requests==2.28.1
rjsmin==1.2.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
